Modify solution

+++ seaoil_flutter/lib/view_models/findus_models.dart
@@ -107,7 +107,7 @@ class StationDetail {
   String stationLat;
   String stationLong;
   String address;
-  double distance;
+  int distance;
 
   StationDetail(
       {required this.stationName,
